requirements.txt
django==1.8
djangorestframework
djangorestframework-xml
djangorestframework-csv
django-filter
django-crispy-forms
django-guardian
